Work around some invalid cookies.
Internally, cookies are parsed by Girder using python's built-in
http.cookies modules (or a backport of that module in Python 2).  This
module is conformant to RFC specifications on cookies.  Some
non-conformant processes can set cookies that do not parse with this
module.  Specifically, in the field we've see cookies like "KEY=part1
part2" with a space in their value.  When the Python module finds
particular types of invalid cookies, it refuses to parse any cookies.
This prevents Girder from working as expected.

This fix allows spaces to be in the cookie values and works around the
immediate problem.  I think this shouldn't cause any risk, and will
allow Girder to work in situations where it would otherwise fail.