Start at the last image in the browser widget.
This uses new browser widget features and works with virtual folders.

This also gets all users and root items, rather than an arbitrary
number.  When there are more users or root items than the arbitrary
limit, there was no indication that not all items were loaded.  Here, an
excessive number of users or collections could break the UI, but this is
unlikely to happen and would at least be an obvious problem instead of
appearing to have missing data.