Prevent invalid overview resizing.
If the overview panel is collapsed and is asked to rerender, the map
ended up being 512 pixels wide.  There were two separate effects that
could result in this behavior.  When a new analysis task is loaded, the
overview panel has the tile source updated, even though it didn't
change.  Any such update resulted in recreating the map unnecessarily.

When the map was created, if the div for the map was not visible due to
being collapsed, it has a zero width.  geojs defaults to 512 pixels wide
in this case.  Further, if the div changed size without the window
changing size, that map was not updated appropriately (for instance,
when the scroll bar appears or disappears).  Use a ResizeObserver, if
available, to adjust the map when either condition occurs.