Add a metadata plot panel.
If metadata exists in the form key: array of objects, and the first
element of such an array has numerical values, then the option to plot
the data, possibly with data from sibling items in the same folder is
presented.  This pulls plotly from a CDN if needed.  If the current
item's metadata doesn't allow for a plot, no extra http requests are
made.

Use closest hovermode in plotly.  The default hovermode is unintuitive.
Add axis titles when the graph is maximized.

Added a control for maximizing the plot size.