Add annotation legends
When annotations have been loaded enough, the first element that has
fill color, line color, or line width information is used to show the
style of that group.  For individual annotations, this is also shown
with the group counts.  As part of this, the count area has been
refactored slightly to prevent long group names from making the marquee
selector tool from becoming unreachable.